requirements.txt
Django==4.1.*
djangorestframework==3.14.*
psycopg2-binary==2.9.*
gunicorn==20.1.*
drf_spectacular==0.28.0
pytest==8.3.4
pytest-django==4.8.0 